requirements.txt
# Core LLM & RAG
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-text-splitters==0.2.4
openai==1.45.0

# Vector Store
faiss-cpu==1.8.0

# Document Loading
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.1

# API & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sse-starlette==2.1.3

# Frontend
streamlit==1.38.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Utilities
pydantic==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
httpx==0.27.2

# Logging & Monitoring
structlog==24.4.0
